+++ package.json
@@ -14,7 +14,7 @@
   },
   "dependencies": {
     "@emotion/react": "^11.14.0",
-    "@emotion/styled": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
     "@tanstack/react-query": "^5.83.0",
     "axios": "^1.10.0",
     "lottie-react": "^2.4.1",
