+++ src/context/AuthContextProvider.tsx
@@ -1,22 +1,12 @@
 import React, { useCallback } from "react"
-import { useMutation, useQueryClient } from "@tanstack/react-query"
-import axios from "axios"
+import { useQueryClient } from "@tanstack/react-query"
 import { AuthContext, AuthContextType } from "./AuthContext"
+import { useLogin } from "@/hooks/useLogin"
 
-const STORAGE_KEYS = {
-  token: "authToken",
-  email: "email",
-  name: "name",
-} as const
-
-interface LoginResponse {
-  code: number
-  data: {
-    authToken: string
-    email: string
-    name: string
-  }
-  error?: string
+interface LoginData {
+  authToken: string
+  email: string
+  name: string
 }
 
 interface AuthContextProviderProps {
@@ -25,29 +15,7 @@ interface AuthContextProviderProps {
 
 export function AuthContextProvider({ children }: AuthContextProviderProps) {
   const qc = useQueryClient()
-
-  const loginMutation = useMutation<
-    LoginResponse["data"],
-    Error,
-    { email: string; password: string }
-  >({
-    mutationFn: async ({ email, password }) => {
-      const baseUrl = import.meta.env.VITE_BASE_URL
-      const loginUrl = new URL("/api/login", baseUrl).toString()
-      const { data } = await axios.post<LoginResponse>(loginUrl, {
-        email,
-        password,
-      })
-      return data.data
-    },
-    onSuccess: (data) => {
-      qc.setQueryData(["auth"], data)
-
-      localStorage.setItem(STORAGE_KEYS.token, data.authToken)
-      localStorage.setItem(STORAGE_KEYS.email, data.email)
-      localStorage.setItem(STORAGE_KEYS.name, data.name)
-    },
-  })
+  const loginMutation = useLogin()
 
   const login = useCallback(
     async (email: string, password: string) => {
@@ -67,20 +35,19 @@ export function AuthContextProvider({ children }: AuthContextProviderProps) {
   const logout = useCallback(() => {
     qc.removeQueries({ queryKey: ["auth"] })
 
-    localStorage.removeItem(STORAGE_KEYS.token)
-    localStorage.removeItem(STORAGE_KEYS.email)
-    localStorage.removeItem(STORAGE_KEYS.name)
+    localStorage.removeItem("authToken")
+    localStorage.removeItem("email")
+    localStorage.removeItem("name")
 
     console.log("logout 완료")
   }, [qc])
 
-  const cached = (qc.getQueryData<LoginResponse["data"]>([
-    "auth",
-  ]) as LoginResponse["data"]) ?? {
-    authToken: localStorage.getItem(STORAGE_KEYS.token),
-    email: localStorage.getItem(STORAGE_KEYS.email) ?? "",
-    name: localStorage.getItem(STORAGE_KEYS.name) ?? "",
+  const cached = (qc.getQueryData<LoginData>(["auth"]) as LoginData) ?? {
+    authToken: localStorage.getItem("authToken"),
+    email: localStorage.getItem("email") ?? "",
+    name: localStorage.getItem("name") ?? "",
   }
+  
   const value: AuthContextType = {
     isLoggedIn: !!cached.authToken,
     authToken: cached.authToken,
