+++ src/functions/fetchHandler.tsx
@@ -1,0 +1,6 @@
+import axios from "axios"
+
+const fetchHandler = <T,>(url: string) =>
+  axios.get<T>(url).then((res) => res.data)
+
+export default fetchHandler
