--- src/hooks/useFetch.tsx
@@ -1,5 +1,0 @@
-import axios from "axios"
-
-const useFetch = <T,>(url: string) => axios.get<T>(url).then((res) => res.data)
-
-export default useFetch
