+++ src/pages/ProductDetail.tsx
@@ -1,0 +1,5 @@
+const ProductDetail = () => {
+  return <div></div>
+}
+
+export default ProductDetail
