--- tsconfig.tsbuildinfo
@@ -1,1 +1,0 @@
-{"root":["./src/app.tsx","./src/home.tsx","./src/emotion.d.ts","./src/main.tsx","./src/mock.ts","./src/mock_card.tsx","./src/mock_present.ts","./src/vite-env.d.ts","./src/components/blank.tsx","./src/components/box.tsx","./src/components/buttonwithimg.tsx","./src/components/buttonwithstyle.tsx","./src/components/card.tsx","./src/components/cardsection.tsx","./src/components/column.tsx","./src/components/giftcardlistlayout.tsx","./src/components/giftcardstyle.tsx","./src/components/grid.tsx","./src/components/indexbadge.tsx","./src/components/inputblank.tsx","./src/components/inputform.tsx","./src/components/layout.tsx","./src/components/loginform.tsx","./src/components/morebutton.tsx","./src/components/navibar.tsx","./src/components/orderlayout.tsx","./src/components/ordermessage.tsx","./src/components/plusbutton.tsx","./src/components/presentcardstyle.tsx","./src/components/presentgiverform.tsx","./src/components/presentlayout.tsx","./src/components/presentreciverform.tsx","./src/components/presentwho.tsx","./src/components/presentwhobackground.tsx","./src/components/productgrid.tsx","./src/components/productimage.tsx","./src/components/row.tsx","./src/components/rowform.tsx","./src/components/screen.tsx","./src/components/sentencetab.tsx","./src/components/sentencesrow.tsx","./src/components/tabbutton.tsx","./src/components/text.tsx","./src/components/themeheader.tsx","./src/components/themeproductsection.tsx","./src/components/title.tsx","./src/components/trending.tsx","./src/components/presentform/addplusbutton.tsx","./src/components/presentform/addreciverbutton.tsx","./src/components/presentform/inputform.tsx","./src/components/presentform/orderform.tsx","./src/components/presentform/presentgiverform.tsx","./src/components/presentform/receiverform.tsx","./src/components/presentform/receiverlist.tsx","./src/components/presentform/receivermodal.tsx","./src/components/presentform/components/makerow.tsx","./src/components/presentform/components/makerowunder.tsx","./src/components/presentform/components/overlay.tsx","./src/components/presentform/components/receivercard.tsx","./src/components/presentform/components/recivercontainer.tsx","./src/components/presenttheme/loading.tsx","./src/components/presenttheme/themenotfound.tsx","./src/constants/messages.tsx","./src/constants/routes.tsx","./src/context/authcontext.tsx","./src/context/authcontextprovider.tsx","./src/context/cardcontext.tsx","./src/context/cardprovider.tsx","./src/context/presentcontext.tsx","./src/context/presentprovider.tsx","./src/context/presentthemecontext.tsx","./src/context/presentthemeprovider.tsx","./src/functions/checkvalid.tsx","./src/functions/fetchthemeproduct.tsx","./src/functions/geterrormessage.tsx","./src/functions/getroute.tsx","./src/hooks/usefetch.tsx","./src/hooks/useinview.ts","./src/hooks/useinfinitescrolling.ts","./src/hooks/useinput.ts","./src/hooks/uselogin.tsx","./src/hooks/useorder.tsx","./src/hooks/useproductinfo.ts","./src/hooks/usequerystate.ts","./src/hooks/usethemeinfo.ts","./src/hooks/usethemeproduct.tsx","./src/interfaces/brandinfo.ts","./src/interfaces/componentstyle.ts","./src/interfaces/price.ts","./src/interfaces/product.ts","./src/interfaces/productresponse.ts","./src/interfaces/productresponsesingle.ts","./src/interfaces/productsummary.ts","./src/interfaces/themeinfo.ts","./src/interfaces/themeinforesponse.ts","./src/interfaces/themeproductlist.tsx","./src/interfaces/themeproductlistresponse.tsx","./src/interfaces/valuetype.ts","./src/pages/buttonsrow.tsx","./src/pages/cardthumbnail.tsx","./src/pages/cardsrow.tsx","./src/pages/categories.tsx","./src/pages/login.tsx","./src/pages/modal.tsx","./src/pages/mypage.tsx","./src/pages/notfound.tsx","./src/pages/orderpage.tsx","./src/pages/plusnewperson.tsx","./src/pages/previewcard.tsx","./src/pages/productinfo.tsx","./src/pages/themeinfopage.tsx","./src/shared/protectedroute.tsx","./src/shared/router.tsx","./src/styles/globalstyle.tsx","./src/styles/reset.tsx","./src/styles/theme.tsx","./src/utils/cookie.tsx"],"version":"5.8.3"}
